release 1.3.1

+++ AuthMe.xcframework/ios-x86_64-simulator/AuthMe.framework/Headers/EngineWrapper.h
@@ -22,7 +22,7 @@ long engine_createSession(long model_);
 
 void engine_releaseSession(long model_, long session_);
 
-float engine_laserTagDetection(long model_, long session_, UIImage* image);
+int engine_laserTagDetection(long model_, long session_, UIImage* image);
 
 void getTransformedLeftPoint(float w,
                              float h,
